requirements.txt
rich==13.7.0
click==8.1.7
pydantic==2.5.0
requests==2.31.0
dnspython==2.4.2
nmap==0.0.1
python-nmap==0.7.1
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.2
paramiko==3.4.0
urllib3==2.1.0
certifi==2023.7.22
cryptography==41.0.7
validators==0.22.0
colorama==0.4.6
tqdm==4.66.1
tabulate==0.9.0
pyyaml==6.0.1
jinja2==3.1.2
sqlalchemy==2.0.23
